requirements.txt
discord.py>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
python-multipart>=0.0.6
